going in a bit of a circle with this struct
