+++ Assets/Scripts/GazeInt/StreamData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c034f3f095a8c92478cb2ae84379ec22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
